Add parallel testing for nightly builds (#224)
Nightly builds are starting to become pretty big, and since the bottleneck is the CPU as opposed to the GPU, we can leverage multiple cores instead of doing everything sequentially. This change distributes the test between 6 processes in a multiprocessing pool in order to keep the CPU and the GPU as busy as possible.