Decouple ABSL dependencies from TF (#247)
Currently, every time we update the TF version, we need to make sure that the version of our Abseil library is the same as the version of TF that we want to be compatible with, which is very bad. This is because we include the Abseil headers from the TF wheel instead of using the headers compatible with our version of Abseil. Decoupling them will allow our wheel to be compatible with far more TF wheels and will mitigate build breaks every time we update the TF version.