Add int32 support for StridedSlice (#252)
We've been having trouble getting our DEVICE_DEFAULT int32 registration for StridedSlice make it into TensorFlow Core, so the workaround is to use the eager API to implement it for the DML device in terms of CPU device. This will fix colocation issues that we've been seeing in the wild, especially for transformer models. This might also increase performance by reducing the possibilities of CPU <-> GPU copies.