Fix FloorMod test that failed because of accuracy (#270)
Division in HLSL is allowed to have an error of 1 ULP, so it's possible that a division that should return an exact result (e.g. 6.0 / 2.0 == 3.0) actually returns an approximation (e.g. 6.0 / 2.0 == 2.99999). In most scenarios this is perfectly fine, but when using floor, it makes a huge difference and the end result will be way off (floor(2.99999) != floor(3.0)). For this reason, we cannot accurately test cases where x / y results in an integer.