Update the DML Descriptor BFC Allocator (#276)
Since we first added an implementation of the BFC allocator for the DML descriptor allocator, the TensorFlow core codebase made changes to their own version, which include getting rid of a memory leak. Even though that memory leak wasn't anything disastrous, getting rid of it will make it easier to analyze real memory leaks in the future.