Fix memory leak in StridedSlice (#280)
The RAII wrapper that we used to deallocate the output of the StridedSlice op that was run on the CPU through the Eager API wasn't actually deallocating the memory because we were passing a pointer that was always null. To fix this, we needed to pass a pointer to the pointer (TFE_TensorHandle**).